Updated home page
Added date pickers on home page to select dates and display weather data between the selected dates. Based on the propability of rain in the next 3 days, a message appears to explain if it will rain or no. Updated DB functions to search for data and update/insert weather data if it already exists or not